Add code to create shift object that is handled when dealing with shift form and add supporting code for it

+++ shift_scheduler_bot/SchedulerFormApp/ShiftClass.cs
@@ -16,17 +16,17 @@ namespace scheduler_test1
         private string monthName;
         private string dayName;
         private int dayNum;
-        private int startHour;
-        private int endHour;
+        private double startHour;
+        private double endHour;
 
-        public int getStartHour() { return startHour; }
-        public int getEndHour() { return endHour; }
+        public double getStartHour() { return startHour; }
+        public double getEndHour() { return endHour; }
         public string getWorkerName() { return workerName; }
         public string getMonthName() {  return monthName; }
         public string getDayName() { return dayName; }
         public int getDayNum() { return dayNum; }
-        public void setStartHour(int hour) { this.startHour = hour; }
-        public void setEndHour(int hour) { this.endHour = hour; }
+        public void setStartHour(double hour) { this.startHour = hour; }
+        public void setEndHour(double hour) { this.endHour = hour; }
         public void setWorkerName(string name) { this.workerName = name; }
         public void setMonthName(string month) { this.monthName = month; }
         public void setDayName(string day) { this.dayName = day; }
